requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via -r requirements.in
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via aiohttp
azure-ai-documentintelligence==1.0.0b4
    # via -r requirements.in
azure-core==1.31.0
    # via
    #   azure-ai-documentintelligence
    #   azure-storage-blob
azure-storage-blob==12.23.1
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==43.0.3
    # via azure-storage-blob
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.4
    # via -r requirements.in
fastapi-cli==0.0.5
    # via fastapi
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via
    #   fastapi
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
isodate==0.7.2
    # via
    #   azure-ai-documentintelligence
    #   azure-storage-blob
jinja2==3.1.4
    # via fastapi
jiter==0.6.1
    # via openai
loguru==0.7.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.52.2
    # via -r requirements.in
pillow==11.0.0
    # via pypdf
propcache==0.2.0
    # via yarl
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   fastapi
    #   openai
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pypdf==5.1.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-magic==0.4.27
    # via -r requirements.in
python-multipart==0.0.16
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   azure-core
    #   tiktoken
rich==13.9.3
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.41.2
    # via fastapi
tiktoken==0.8.0
    # via -r requirements.in
tqdm==4.66.6
    # via openai
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   azure-ai-documentintelligence
    #   azure-core
    #   azure-storage-blob
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.3
    # via requests
uvicorn==0.32.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
yarl==1.17.0
    # via aiohttp
